roastpet_cli/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "roastpet-cli"
version = "0.1.0"
description = "RoastPet CLI: AI code roasting companion with RPG mechanics."
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
dependencies = ["watchdog", "rich", "python-dotenv", "openai", "pygame"]

[project.urls]
"Homepage" = "https://github.com/your-org/roastpet"

[tool.setuptools.packages.find]
where = ["."]

[project.scripts]
roastpet_cli = "roastpet_cli.cli:main"
